pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-mcp-azure-oauth"
version = "1.0.0"
description = "RFC-compliant Azure AD OAuth 2.0 router for FastAPI MCP servers with Copilot Studio support"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Lee Pasifull", email = "lee@pasifull.com" }]
requires-python = ">=3.10"
keywords = [
    "fastapi",
    "azure",
    "oauth",
    "oauth2",
    "mcp",
    "model-context-protocol",
    "copilot-studio",
    "authentication",
    "jwt",
    "azure-ad",
    "entra-id",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "PyJWT[crypto]>=2.8.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.0",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/LeeP-Tech/fastapi-mcp-azure-oauth"
Repository = "https://github.com/LeeP-Tech/fastapi-mcp-azure-oauth"
"Bug Tracker" = "https://github.com/LeeP-Tech/fastapi-mcp-azure-oauth/issues"
Changelog = "https://github.com/LeeP-Tech/fastapi-mcp-azure-oauth/blob/main/CHANGELOG.md"
Documentation = "https://github.com/LeeP-Tech/fastapi-mcp-azure-oauth#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_mcp_azure_oauth"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short"

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
